1.fix typo 2.move require to top 3.using string template 4.move logger to a separate js file 5.change to absolute path
