+++ src/sync/index.js
@@ -3,40 +3,7 @@
 const _ = require('lodash');
 const { Qweb3, Contract } = require('qweb3');
 
-
-
-const loglvl = process.env.loglvl || 'info';
-
-var winston = require('winston');
-var winston_config = winston.config;
-var logger = new (winston.Logger)({
-  transports: [
-    new (winston.transports.Console)({
-      timestamp: function() {
-        return new Date().toISOString().slice(0,19); 
-      },
-      formatter: function(options) {
-        return options.timestamp() +' '+ __filename+' ' +
-          winston_config.colorize(options.level, options.level.toUpperCase()) + ' ' +
-          (options.message ? options.message : '') +
-          (options.meta && Object.keys(options.meta).length ? '\n\t'+ JSON.stringify(options.meta) : '' );
-      }
-    }),
-    new (winston.transports.File)({
-      filename: './logs/'+new Date().toISOString().slice(0,10)+'.log', 
-      
-    })
-  ]
-});
-
-
-
-logger.level = loglvl;
-logger.on('error', function(err) {
-  if (err.code == 'ENOENT') {
-    mkdirp.sync(path.join(__dirname, 'logs'));
-  }
-});
+const logger = require('../utils/logger');
 
 const config = require('../config/config');
 const connectDB = require('../db/nedb');
